Update of my profile page
Made the buttons a bit smaller since they were too big. Change one of the textbox for a spinner and added another spinner. This spinners will be one for the user to choose their dietary and other for the user to choose if they want their profile has public or private